endianness got me all paranoid about using int... so now I'm using uint8_t, uint16_t, uint32_t pretty much everywhere.

+++ src/core/framework/network/client/NetworkManagerClient.h
@@ -62,7 +62,7 @@ public:
     
     void requestToAddLocalPlayer();
     
-    void requestToDropLocalPlayer(int index);
+    void requestToDropLocalPlayer(uint8_t index);
     
     const WeightedTimedMovingAverage& getBytesReceivedPerSecond() const;
     
@@ -107,7 +107,7 @@ private:
     WeightedTimedMovingAverage* m_avgRoundTripTime;
     
     bool m_isRequestingToAddLocalPlayer;
-    int m_isRequestingToDropLocalPlayer;
+    uint8_t m_isRequestingToDropLocalPlayer;
     
     void processPacket(InputMemoryBitStream& inInputStream, IMachineAddress* inFromAddress);
