+++ src/core/game/network/Server.h
@@ -18,19 +18,19 @@ class Robot;
 class Server
 {
 public:
-    static void create(bool isSteam, int inNumCratesToSpawn = 6, int inNumSpacePiratesToSpawn = 4);
+    static void create(bool isSteam, uint32_t inNumCratesToSpawn = 6, uint32_t inNumSpacePiratesToSpawn = 4);
     
     static void destroy();
     
     static Server* getInstance();
     
-    static void sHandleNewClient(int playerId, std::string playerName);
+    static void sHandleNewClient(uint8_t playerId, std::string playerName);
     
-    static void sHandleLostClient(ClientProxy* inClientProxy, int index);
+    static void sHandleLostClient(ClientProxy* inClientProxy, uint8_t index);
     
     void update(float deltaTime);
     
-    int getPlayerIdForRobotBeingCreated();
+    uint8_t getPlayerIdForRobotBeingCreated();
     
     void toggleEnemies();
     
@@ -50,20 +50,20 @@ private:
     float m_fStateTime;
     float m_fFrameStateTime;
     float m_fStateTimeNoEnemies;
-    int m_iPlayerIdForRobotBeingCreated;
-    int m_iNumCratesToSpawn;
-    int m_iNumSpacePiratesToSpawn;
+    uint8_t m_iPlayerIdForRobotBeingCreated;
+    uint32_t m_iNumCratesToSpawn;
+    uint32_t m_iNumSpacePiratesToSpawn;
     bool m_isSpawningEnemies;
     bool m_isSpawningObjects;
     bool m_isDisplaying;
     
-    void handleNewClient(int playerId, std::string playerName);
+    void handleNewClient(uint8_t playerId, std::string playerName);
     
-    void handleLostClient(ClientProxy* inClientProxy, int index);
+    void handleLostClient(ClientProxy* inClientProxy, uint8_t index);
     
     void deleteRobotWithPlayerId(uint8_t playerId);
     
-    void spawnRobotForPlayer(int inPlayerId, std::string inPlayerName);
+    void spawnRobotForPlayer(uint8_t inPlayerId, std::string inPlayerName);
     
     void respawnEnemiesIfNecessary();
     
@@ -72,7 +72,7 @@ private:
     void clearClientMoves();
     
     // ctor, copy ctor, and assignment should be private in a Singleton
-    Server(bool isSteam, int inNumCratesToSpawn, int inNumSpacePiratesToSpawn);
+    Server(bool isSteam, uint32_t inNumCratesToSpawn, uint32_t inNumSpacePiratesToSpawn);
     ~Server();
     Server(const Server&);
     Server& operator=(const Server&);
